Create MapResource
